requirements.txt
discord.py==0.16.8
PyNaCl==1.0.1
beautifulsoup4==4.6.0
lxml==3.8.0
requests==2.18.1
async-timeout==1.2.1
pymongo==3.4.0